@@ -1,20 +1,16 @@
-const chai = require('chai')
-const IdeasLoader = require('../app/utils/ideasLoader')
-const Shuffled = require('../app/utils/shuffled')
-
-chai.should()
+import 'chai/register-should'
+import IdeasLoader from '../app/utils/ideasLoader'
+import Shuffled from '../app/utils/shuffled'
 
 describe('ideasLoader', function () {
-  beforeEach(function () {
-    this.ideas = new IdeasLoader([
-      { data: 'a', enabled: true },
-      { data: 'b', enabled: false },
-      { data: 'c', enabled: true }
-    ])
-  })
+  const ideas = new IdeasLoader([
+    { data: 'a', enabled: true },
+    { data: 'b', enabled: false },
+    { data: 'c', enabled: true }
+  ])
 
-  it('returns enabled ideas', function () {
+  it('returns enabled ideas', () => {
     const enabled = new Shuffled(['a', 'c'])
-    this.ideas.ideas().should.be.deep.equal(enabled)
+    ideas.ideas().should.be.deep.equal(enabled)
   })
 })
