@@ -1,9 +1,9 @@
-const chai = require('chai')
-const path = require('path')
-const NaturalBreaksManager = require('../app/utils/naturalBreaksManager')
-const Store = require('electron-store')
-
-chai.should()
+import 'chai/register-should'
+import { join } from 'path'
+import NaturalBreaksManager from '../app/utils/naturalBreaksManager'
+import Store from 'electron-store'
+import defaultSettings from '../app/utils/defaultSettings'
+import { unlink } from 'fs'
 
 describe('naturalBreaksManager', function () {
   let settings = null
@@ -11,9 +11,9 @@ describe('naturalBreaksManager', function () {
 
   beforeEach(() => {
     settings = new Store({
-      cwd: path.join(__dirname),
-      name: 'test-settings',
-      defaults: require('../app/utils/defaultSettings')
+      cwd: join(__dirname),
+      name: 'test-settings-naturalBreaksManager',
+      defaults: defaultSettings
     })
     naturalBreaksManager = new NaturalBreaksManager(settings)
   })
@@ -65,7 +65,7 @@ describe('naturalBreaksManager', function () {
     naturalBreaksManager = null
 
     if (settings) {
-      require('fs').unlink(path.join(__dirname, '/test-settings.json'), (_) => {})
+      unlink(join(__dirname, '/test-settings-naturalBreaksManager.json'), (_) => {})
       settings = null
     }
   })
