@@ -6,8 +6,9 @@ const {
 const path = require('path')
 const i18next = require('i18next')
 const Backend = require('i18next-fs-backend')
-const log = require('electron-log')
+const log = require('electron-log/main')
 const Store = require('electron-store')
+const { registerBreakShortcuts } = require('./utils/breakShortcuts')
 
 process.on('uncaughtException', (err, _) => {
   log.error(err)
@@ -29,7 +30,7 @@ nativeTheme.on('updated', function theThemeHasChanged () {
   if (!gotTheLock) {
     return
   }
-  appIcon.setImage(trayIconPath())
+  updateTray()
 })
 
 const Utils = require('./utils/utils')
@@ -37,12 +38,14 @@ const IdeasLoader = require('./utils/ideasLoader')
 const BreaksPlanner = require('./breaksPlanner')
 const AppIcon = require('./utils/appIcon')
 const { UntilMorning } = require('./utils/untilMorning')
+const AutostartManager = require('./utils/autostartManager')
 const Command = require('./utils/commands')
 
 let microbreakIdeas
 let breakIdeas
 let breakPlanner
 let appIcon = null
+let autostartManager = null
 let processWin = null
 let microbreakWins = null
 let breakWins = null
@@ -54,10 +57,13 @@ let myStretchlyWindow = null
 let settings
 let pausedForSuspendOrLock = false
 let nextIdea = null
-let appIsQuitting = false
 let updateChecker
+let currentTrayIconPath = null
+let currentTrayMenuTemplate = null
+let trayUpdateIntervalObj = null
 
 require('@electron/remote/main').initialize()
+log.initialize({ preload: true })
 
 app.setAppUserModelId('net.hovancik.stretchly')
 
@@ -141,6 +147,11 @@ if (!gotTheLock) {
         pauseBreaks(duration)
         break
       }
+
+      case 'preferences':
+        log.info('Stretchly: open Preferences window (requested by second instance)')
+        createPreferencesWindow()
+        break
     }
   })
 }
@@ -149,12 +160,19 @@ app.on('ready', initialize)
 app.on('window-all-closed', () => {
   // do nothing, so app wont get closed
 })
-app.on('before-quit', () => {
-  appIsQuitting = true
-  globalShortcut.unregisterAll()
+app.on('before-quit', (event) => {
+  if ((breakPlanner.scheduler.reference === 'finishMicrobreak' && settings.get('microbreakStrictMode')) ||
+      (breakPlanner.scheduler.reference === 'finishBreak' && settings.get('breakStrictMode'))
+  ) {
+    log.info('Stretchly: preventing app closure (in break with strict mode)')
+    event.preventDefault()
+  } else {
+    globalShortcut.unregisterAll()
+    app.quit()
+  }
 })
 
-function initialize (isAppStart = true) {
+async function initialize (isAppStart = true) {
   if (!gotTheLock) {
     return
   }
@@ -162,7 +180,39 @@ function initialize (isAppStart = true) {
   log.info(`Stretchly: ${isAppStart ? '' : 're'}initializing...`)
   require('events').defaultMaxListeners = 200 // for watching Store changes
   if (!settings) {
-    settings = new Store({ defaults: require('./utils/defaultSettings'), watch: true })
+    settings = new Store({
+      defaults: require('./utils/defaultSettings'),
+      beforeEachMigration: (store, context) => {
+        log.info(`Stretchly: migrating preferences from Stretchly v${context.fromVersion} to v${context.toVersion}`)
+      },
+      migrations: {
+        '1.13.0': store => {
+          if (store.has('pauseBreaksShortcut')) {
+            store.set('pauseBreaksToggleShortcut', store.get('pauseBreaksShortcut'))
+            log.info(`Stretchly: settings pauseBreaksToggleShortcut to "${store.get('pauseBreaksShortcut')}"`)
+            store.delete('pauseBreaksShortcut')
+            log.info('Stretchly: removing pauseBreaksShortcut')
+          } else {
+            log.info('Stretchly: not migrating pauseBreaksShortcut')
+          }
+          if (store.has('pauseBreaksShortcut')) {
+            store.delete('resumeBreaksShortcut')
+            log.info('Stretchly: removing resumeBreaksShortcut')
+          }
+        },
+        '1.17.0': store => {
+          if (store.has('showBreakActionsInStrictMode')) {
+            store.set('showTrayMenuInStrictMode', store.get('showBreakActionsInStrictMode'))
+            log.info(`Stretchly: settings showTrayMenuInStrictMode to "${store.get('showBreakActionsInStrictMode')}"`)
+            store.delete('showBreakActionsInStrictMode')
+            log.info('Stretchly: removing showBreakActionsInStrictMode')
+          } else {
+            log.info('Stretchly: not migrating showBreakActionsInStrictMode')
+          }
+        }
+      },
+      watch: true
+    })
     log.info('Stretchly: loading preferences')
     Store.initRenderer()
     Object.entries(settings.store).forEach(([key, _]) => {
@@ -191,12 +241,12 @@ function initialize (isAppStart = true) {
     breakPlanner.naturalBreaks(settings.get('naturalBreaks'))
     breakPlanner.nextBreak()
   }
-  if (!appIcon) {
-    if (process.platform === 'darwin') {
-      app.dock.hide()
-    }
-    appIcon = new Tray(trayIconPath())
-  }
+
+  autostartManager = new AutostartManager({
+    platform: process.platform,
+    windowsStore: process.windowsStore,
+    app
+  })
 
   startI18next()
   startProcessWin()
@@ -219,38 +269,26 @@ function initialize (isAppStart = true) {
   })
   startPowerMonitoring()
   if (preferencesWin) {
-    preferencesWin.send('renderSettings', settingsToSend())
+    preferencesWin.send('renderSettings', await settingsToSend())
   }
   if (welcomeWin) {
-    welcomeWin.send('renderSettings', settingsToSend())
+    welcomeWin.send('renderSettings', await settingsToSend())
   }
   if (contributorPreferencesWindow) {
-    contributorPreferencesWindow.send('renderSettings', settingsToSend())
+    contributorPreferencesWindow.send('renderSettings', await settingsToSend())
   }
   globalShortcut.unregisterAll()
-  if (settings.get('resumeBreaksShortcut') !== '') {
-    const resumeBreaksShortcut = globalShortcut.register(settings.get('resumeBreaksShortcut'), () => {
-      resumeBreaks(false)
-    })
 
-    if (!resumeBreaksShortcut) {
-      log.warn('Stretchly: resumeBreaksShortcut registration failed')
-    } else {
-      log.info(`Stretchly: resumeBreaksShortcut registration successful (${settings.get('resumeBreaksShortcut')})`)
-    }
-  }
-  if (settings.get('pauseBreaksShortcut') !== '') {
-    const pauseBreaksShortcut = globalShortcut.register(settings.get('pauseBreaksShortcut'), () => {
-      pauseBreaks(1)
-    })
+  registerBreakShortcuts({
+    settings,
+    log,
+    globalShortcut,
+    breakPlanner,
+    functions: { pauseBreaks, resumeBreaks, skipToBreak, skipToMicrobreak, resetBreaks }
+  })
 
-    if (!pauseBreaksShortcut) {
-      log.warn('Stretchly: pauseBreaksShortcut registration failed')
-    } else {
-      log.info(`Stretchly: pauseBreaksShortcut registration successful (${settings.get('pauseBreaksShortcut')})`)
-    }
-  }
   loadIdeas()
+  updateTray()
 }
 
 function startI18next () {
@@ -268,17 +306,12 @@ function startI18next () {
       if (err) {
         console.log(err.stack)
       }
-      if (appIcon) {
-        updateTray()
-        setInterval(updateTray, 10000)
-      }
+      updateTray()
     })
 }
 
 i18next.on('languageChanged', function (lng) {
-  if (appIcon) {
-    updateTray()
-  }
+  updateTray()
   loadIdeas()
 })
 
@@ -462,7 +495,7 @@ function trayIconPath () {
     darkMode: nativeTheme.shouldUseDarkColors,
     platform: process.platform,
     timeToBreakInTray: settings.get('timeToBreakInTray'),
-    timeToBreak: Utils.minutesRemaining(breakPlanner.scheduler.timeLeft),
+    timeToBreak: Utils.minutesRemaining(breakPlanner.timeToNextBreak),
     reference: breakPlanner.scheduler.reference
   }
   const trayIconFileName = new AppIcon(params).trayIconFileName
@@ -496,7 +529,8 @@ function startProcessWin () {
     backgroundThrottling: false,
     webPreferences: {
       preload: path.join(__dirname, './process.js'),
-      enableRemoteModule: true
+      enableRemoteModule: true,
+      sandbox: false
     }
   })
   require('@electron/remote/main').enable(processWin.webContents)
@@ -511,15 +545,16 @@ function createWelcomeWindow (isAppStart = true) {
     const modalPath = path.join('file://', __dirname, '/welcome.html')
     welcomeWin = new BrowserWindow({
       x: displaysX(-1, 1000),
-      y: displaysY(-1, 770),
+      y: displaysY(-1, 950),
       width: 1000,
-      height: 770,
+      height: 950,
       autoHideMenuBar: true,
       icon: windowIconPath(),
       backgroundColor: 'EDEDED',
       webPreferences: {
         preload: path.join(__dirname, './welcome.js'),
-        enableRemoteModule: true
+        enableRemoteModule: true,
+        sandbox: false
       }
     })
     require('@electron/remote/main').enable(welcomeWin.webContents)
@@ -550,7 +585,8 @@ function createContributorSettingsWindow () {
     backgroundColor: 'EDEDED',
     webPreferences: {
       preload: path.join(__dirname, './contributor-preferences.js'),
-      enableRemoteModule: true
+      enableRemoteModule: true,
+      sandbox: false
     }
   })
   require('@electron/remote/main').enable(contributorPreferencesWindow.webContents)
@@ -582,7 +618,8 @@ function createSyncPreferencesWindow () {
     backgroundColor: 'whitesmoke',
     webPreferences: {
       preload: path.resolve(__dirname, './electron-bridge.js'),
-      enableRemoteModule: true
+      enableRemoteModule: true,
+      sandbox: false
     }
   })
   require('@electron/remote/main').enable(syncPreferencesWindow.webContents)
@@ -631,6 +668,22 @@ function startBreakNotification () {
   updateTray()
 }
 
+function getBlurredBackgroundWindowOptions () {
+  if (!settings.get('blurredBackground')) {
+    return {}
+  }
+
+  switch (process.platform) {
+    case 'darwin':
+      return {
+        vibrancy: 'hud',
+        visualEffectState: 'active'
+      }
+    default:
+      return {}
+  }
+}
+
 function startMicrobreak () {
   // don't start another break if break running
   if (microbreakWins) {
@@ -653,7 +706,7 @@ function startMicrobreak () {
   nextIdea = null
 
   if (settings.get('microbreakStartSoundPlaying') && !settings.get('silentNotifications')) {
-    processWin.webContents.send('playSound', settings.get('audio'), settings.get('volume'))
+    processWin.webContents.send('playSound', settings.get('miniBreakAudio'), settings.get('volume'))
   }
 
   for (let localDisplayId = 0; localDisplayId < numberOfDisplays(); localDisplayId++) {
@@ -666,15 +719,20 @@ function startMicrobreak () {
       frame: showBreaksAsRegularWindows,
       show: false,
       backgroundThrottling: false,
-      transparent: settings.get('transparentMode'),
-      backgroundColor: calculateBackgroundColor(),
+      transparent: true,
+      ...getBlurredBackgroundWindowOptions(),
+      backgroundColor: calculateBackgroundColor(settings.get('miniBreakColor')),
       skipTaskbar: !showBreaksAsRegularWindows,
       focusable: showBreaksAsRegularWindows,
       alwaysOnTop: !showBreaksAsRegularWindows,
+      hasShadow: false,
       title: 'Stretchly',
+      titleBarStyle: 'hidden',
+      titleBarOverlay: false,
       webPreferences: {
         preload: path.join(__dirname, './microbreak.js'),
-        enableRemoteModule: true
+        enableRemoteModule: true,
+        sandbox: false
       }
     }
 
@@ -691,6 +749,7 @@ function startMicrobreak () {
     let microbreakWinLocal = new BrowserWindow(windowOptions)
     // seems to help with multiple-displays problems
     microbreakWinLocal.setSize(windowOptions.width, windowOptions.height)
+
     ipcMain.on('send-microbreak-data', (event) => {
       const startTime = Date.now()
       if (!strictMode || postponable) {
@@ -707,10 +766,15 @@ function startMicrobreak () {
       }
       event.sender.send('microbreakIdea', idea)
       event.sender.send('progress', startTime,
-        breakDuration, strictMode, postponable, postponableDurationPercent, calculateBackgroundColor())
+        breakDuration, strictMode, postponable, postponableDurationPercent, calculateBackgroundColor(settings.get('miniBreakColor')))
     })
     // microbreakWinLocal.webContents.openDevTools()
     microbreakWinLocal.once('ready-to-show', () => {
+      log.info('Stretchly: ready-to-show fired')
+    })
+
+    ipcMain.once('mini-break-loaded', () => {
+      log.info('Stretchly: Mini Break window loaded')
       if (showBreaksAsRegularWindows) {
         microbreakWinLocal.show()
       } else {
@@ -722,6 +786,8 @@ function startMicrobreak () {
         if (showBreaksAsRegularWindows) {
           microbreakWinLocal.setFullScreen(settings.get('fullscreen'))
         } else {
+          microbreakWinLocal.setMinimizable(false)
+          microbreakWinLocal.setClosable(false)
           microbreakWinLocal.setKiosk(settings.get('fullscreen'))
         }
       }
@@ -734,6 +800,7 @@ function startMicrobreak () {
           microbreakWinLocal.center()
         }, 0)
       }
+      updateTray()
     })
 
     require('@electron/remote/main').enable(microbreakWinLocal.webContents)
@@ -742,10 +809,10 @@ function startMicrobreak () {
     microbreakWinLocal.setAlwaysOnTop(!showBreaksAsRegularWindows, 'pop-up-menu')
     if (microbreakWinLocal) {
       microbreakWinLocal.on('close', (e) => {
-        if (settings.get('showBreaksAsRegularWindows')) {
-          if (!appIsQuitting && !microbreakWinLocal.fullScreen) {
-            e.preventDefault()
-          }
+        if (breakPlanner.scheduler.timeLeft > 0 && settings.get('microbreakStrictMode')) {
+          // FIXME this will still log when postponing break
+          log.info('Stretchly: preventing closing break window as in strict mode')
+          e.preventDefault()
         }
       })
       microbreakWinLocal.on('closed', () => {
@@ -762,11 +829,10 @@ function startMicrobreak () {
     }
   }
   if (process.platform === 'darwin') {
-    app.dock.hide()
+    if (app.dock.isVisible) {
+      app.dock.hide()
+    }
   }
-  ipcMain.on('mini-break-loaded', (event) => {
-    updateTray()
-  })
 }
 
 function startBreak () {
@@ -804,15 +870,20 @@ function startBreak () {
       frame: showBreaksAsRegularWindows,
       show: false,
       backgroundThrottling: false,
-      transparent: settings.get('transparentMode'),
-      backgroundColor: calculateBackgroundColor(),
+      transparent: true,
+      ...getBlurredBackgroundWindowOptions(),
+      backgroundColor: calculateBackgroundColor(settings.get('mainColor')),
       skipTaskbar: !showBreaksAsRegularWindows,
       focusable: showBreaksAsRegularWindows,
       alwaysOnTop: !showBreaksAsRegularWindows,
+      hasShadow: false,
       title: 'Stretchly',
+      titleBarStyle: 'hidden',
+      titleBarOverlay: false,
       webPreferences: {
         preload: path.join(__dirname, './break.js'),
-        enableRemoteModule: true
+        enableRemoteModule: true,
+        sandbox: false
       }
     }
 
@@ -845,10 +916,15 @@ function startBreak () {
       }
       event.sender.send('breakIdea', idea)
       event.sender.send('progress', startTime,
-        breakDuration, strictMode, postponable, postponableDurationPercent, calculateBackgroundColor())
+        breakDuration, strictMode, postponable, postponableDurationPercent, calculateBackgroundColor(settings.get('mainColor')))
     })
     // breakWinLocal.webContents.openDevTools()
     breakWinLocal.once('ready-to-show', () => {
+      log.info('Stretchly: ready-to-show fired')
+    })
+
+    ipcMain.once('long-break-loaded', () => {
+      log.info('Stretchly: Long Break window loaded')
       if (showBreaksAsRegularWindows) {
         breakWinLocal.show()
       } else {
@@ -860,6 +936,8 @@ function startBreak () {
         if (showBreaksAsRegularWindows) {
           breakWinLocal.setFullScreen(settings.get('fullscreen'))
         } else {
+          breakWinLocal.setMinimizable(false)
+          breakWinLocal.setClosable(false)
           breakWinLocal.setKiosk(settings.get('fullscreen'))
         }
       }
@@ -873,6 +951,7 @@ function startBreak () {
           breakWinLocal.center()
         }, 0)
       }
+      updateTray()
     })
 
     require('@electron/remote/main').enable(breakWinLocal.webContents)
@@ -881,10 +960,10 @@ function startBreak () {
     breakWinLocal.setAlwaysOnTop(!showBreaksAsRegularWindows, 'pop-up-menu')
     if (breakWinLocal) {
       breakWinLocal.on('close', (e) => {
-        if (settings.get('showBreaksAsRegularWindows')) {
-          if (!appIsQuitting && !breakWinLocal.fullScreen) {
-            e.preventDefault()
-          }
+        if (breakPlanner.scheduler.timeLeft > 0 && settings.get('breakStrictMode')) {
+          // FIXME this will still log when postponing break
+          log.info('Stretchly: preventing closing break window as in strict mode')
+          e.preventDefault()
         }
       })
       breakWinLocal.on('closed', () => {
@@ -901,19 +980,19 @@ function startBreak () {
     }
   }
   if (process.platform === 'darwin') {
-    app.dock.hide()
+    if (app.dock.isVisible) {
+      app.dock.hide()
+    }
   }
-  ipcMain.on('long-break-loaded', (event) => {
-    updateTray()
-  })
 }
 
-function breakComplete (shouldPlaySound, windows) {
-  if (globalShortcut.isRegistered(settings.get('endBreakShortcut'))) {
+function breakComplete (shouldPlaySound, windows, breakType) {
+  if (settings.get('endBreakShortcut') && globalShortcut.isRegistered(settings.get('endBreakShortcut'))) {
     globalShortcut.unregister(settings.get('endBreakShortcut'))
   }
   if (shouldPlaySound && !settings.get('silentNotifications')) {
-    processWin.webContents.send('playSound', settings.get('audio'), settings.get('volume'))
+    const audio = breakType === 'mini' ? 'miniBreakAudio' : 'audio'
+    processWin.webContents.send('playSound', settings.get(audio), settings.get('volume'))
   }
   if (process.platform === 'darwin') {
     // get focus on the last app
@@ -923,7 +1002,7 @@ function breakComplete (shouldPlaySound, windows) {
 }
 
 function finishMicrobreak (shouldPlaySound = true, shouldPlanNext = true) {
-  microbreakWins = breakComplete(shouldPlaySound, microbreakWins)
+  microbreakWins = breakComplete(shouldPlaySound, microbreakWins, 'mini')
   log.info(`Stretchly: finishing Mini Break (shouldPlanNext: ${shouldPlanNext})`)
   if (shouldPlanNext) {
     breakPlanner.nextBreak()
@@ -934,7 +1013,7 @@ function finishMicrobreak (shouldPlaySound = true, shouldPlanNext = true) {
 }
 
 function finishBreak (shouldPlaySound = true, shouldPlanNext = true) {
-  breakWins = breakComplete(shouldPlaySound, breakWins)
+  breakWins = breakComplete(shouldPlaySound, breakWins, 'long')
   log.info(`Stretchly: finishing Long Break (shouldPlanNext: ${shouldPlanNext})`)
   if (shouldPlanNext) {
     breakPlanner.nextBreak()
@@ -1002,8 +1081,12 @@ function resetBreaks () {
   updateTray()
 }
 
-function calculateBackgroundColor () {
-  return settings.get('mainColor') + Math.round(settings.get('opacity') * 255).toString(16)
+function calculateBackgroundColor (color) {
+  let opacityMultiplier = 1
+  if (settings.get('transparentMode')) {
+    opacityMultiplier = settings.get('opacity')
+  }
+  return color + Math.round(opacityMultiplier * 255).toString(16).padStart(2, '0')
 }
 
 function loadIdeas () {
@@ -1079,7 +1162,8 @@ function createPreferencesWindow () {
     backgroundColor: '#EDEDED',
     webPreferences: {
       preload: path.join(__dirname, './preferences.js'),
-      enableRemoteModule: true
+      enableRemoteModule: true,
+      sandbox: false
     }
   })
   require('@electron/remote/main').enable(preferencesWin.webContents)
@@ -1093,12 +1177,48 @@ function createPreferencesWindow () {
 }
 
 function updateTray () {
-  updateToolTip()
-  appIcon.setImage(trayIconPath())
-  appIcon.setContextMenu(getTrayMenu())
+  if (process.platform === 'darwin') {
+    if (app.dock.isVisible) {
+      app.dock.hide()
+    }
+  }
+
+  if (!appIcon && !settings.get('showTrayIcon')) {
+    return
+  }
+
+  if (settings.get('showTrayIcon')) {
+    if (!appIcon) {
+      appIcon = new Tray(trayIconPath())
+      appIcon.on('double-click', () => {
+        createPreferencesWindow()
+      })
+      appIcon.on('click', () => {
+        appIcon.popUpContextMenu(Menu.buildFromTemplate(currentTrayMenuTemplate))
+      })
+    }
+    if (!trayUpdateIntervalObj) {
+      trayUpdateIntervalObj = setInterval(updateTray, 10000)
+    }
+
+    updateToolTip()
+
+    const newTrayIconPath = trayIconPath()
+    if (newTrayIconPath !== currentTrayIconPath) {
+      appIcon.setImage(newTrayIconPath)
+      currentTrayIconPath = newTrayIconPath
+    }
+
+    const newTrayMenuTemplate = getTrayMenuTemplate()
+    if (JSON.stringify(newTrayMenuTemplate) !== JSON.stringify(currentTrayMenuTemplate)) {
+      const trayMenu = Menu.buildFromTemplate(newTrayMenuTemplate)
+      appIcon.setContextMenu(trayMenu)
+      currentTrayMenuTemplate = newTrayMenuTemplate
+    }
+  }
 }
 
-function getTrayMenu () {
+function getTrayMenuTemplate () {
   const trayMenu = []
 
   if (global.shared.isNewVersion) {
@@ -1114,8 +1234,8 @@ function getTrayMenu () {
 
   const StatusMessages = require('./utils/statusMessages')
   const statusMessage = new StatusMessages({
-    breakPlanner: breakPlanner,
-    settings: settings
+    breakPlanner,
+    settings
   }).trayMessage
 
   if (statusMessage !== '') {
@@ -1130,6 +1250,15 @@ function getTrayMenu () {
     trayMenu.push({
       type: 'separator'
     })
+  }
+
+  if ((breakPlanner.scheduler.reference === 'finishMicrobreak' && settings.get('microbreakStrictMode') &&
+        !settings.get('showTrayMenuInStrictMode')) ||
+      (breakPlanner.scheduler.reference === 'finishBreak' && settings.get('breakStrictMode') &&
+      !settings.get('showTrayMenuInStrictMode'))
+  ) {
+    // empty menu, we are in strict mode
+    return trayMenu
   }
 
   if (!(breakPlanner.isPaused || breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared)) {
@@ -1149,7 +1278,7 @@ function getTrayMenu () {
     if (settings.get('break') || settings.get('microbreak')) {
       trayMenu.push({
         label: i18next.t('main.skipToTheNext'),
-        submenu: submenu
+        submenu
       })
     }
   }
@@ -1162,36 +1291,37 @@ function getTrayMenu () {
         updateTray()
       }
     })
-  } else if (breakPlanner.scheduler.reference === 'finishMicrobreak' && settings.get('microbreakStrictMode')) {
-    // nothing
-  } else if (breakPlanner.scheduler.reference === 'finishBreak' && settings.get('breakStrictMode')) {
-    // nothing
   } else if (!(breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared)) {
     trayMenu.push({
       label: i18next.t('main.pause'),
       submenu: [
         {
           label: i18next.t('utils.minutes', { count: 30 }),
+          accelerator: settings.get('pauseBreaksFor30MinutesShortcut') || null,
           click: function () {
             pauseBreaks(1800 * 1000)
           }
         }, {
           label: i18next.t('main.forHour'),
+          accelerator: settings.get('pauseBreaksFor1HourShortcut') || null,
           click: function () {
             pauseBreaks(3600 * 1000)
           }
         }, {
           label: i18next.t('main.for2Hours'),
+          accelerator: settings.get('pauseBreaksFor2HoursShortcut') || null,
           click: function () {
             pauseBreaks(3600 * 2 * 1000)
           }
         }, {
           label: i18next.t('main.for5Hours'),
+          accelerator: settings.get('pauseBreaksFor5HoursShortcut') || null,
           click: function () {
             pauseBreaks(3600 * 5 * 1000)
           }
         }, {
           label: i18next.t('main.untilMorning'),
+          accelerator: settings.get('pauseBreaksUntilMorningShortcut') || null,
           click: function () {
             const untilMorning = new UntilMorning(settings).msToSunrise()
             pauseBreaks(untilMorning)
@@ -1244,25 +1374,27 @@ function getTrayMenu () {
     }
   })
 
-  return Menu.buildFromTemplate(trayMenu)
+  return trayMenu
 }
 
 function updateToolTip () {
   const StatusMessages = require('./utils/statusMessages')
   let trayMessage = i18next.t('main.toolTipHeader')
   const message = new StatusMessages({
-    breakPlanner: breakPlanner,
-    settings: settings
+    breakPlanner,
+    settings
   }).trayMessage
   if (message !== '') {
     trayMessage += '\n\n' + message
   }
-  appIcon.setToolTip(trayMessage)
+  if (appIcon) {
+    appIcon.setToolTip(trayMessage)
+  }
 }
 
 function showNotification (text) {
   processWin.webContents.send('showNotification', {
-    text: text,
+    text,
     silent: settings.get('silentNotifications')
   })
 }
@@ -1300,8 +1432,28 @@ ipcMain.on('save-setting', function (event, key, value) {
     nativeTheme.themeSource = value
   }
 
+  if (key === 'audio') {
+    settings.set('miniBreakAudio', value)
+  }
+
+  if (key === 'mainColor') {
+    settings.set('miniBreakColor', value)
+  }
+
+  if (key === 'showTrayIcon') {
+    settings.set('showTrayIcon', value)
+    if (value) {
+      updateTray()
+    } else {
+      clearInterval(trayUpdateIntervalObj)
+      trayUpdateIntervalObj = null
+      appIcon.destroy()
+      appIcon = null
+    }
+  }
+
   if (key === 'openAtLogin') {
-    app.setLoginItemSettings({ openAtLogin: value })
+    autostartManager.setAutostartEnabled(value)
   } else {
     settings.set(key, value)
   }
@@ -1320,24 +1472,22 @@ ipcMain.on('restore-defaults', (event) => {
     message: i18next.t('main.warning'),
     buttons: [i18next.t('main.continue'), i18next.t('main.cancel')]
   }
-  dialog.showMessageBox(dialogOpts).then((returnValue) => {
+  dialog.showMessageBox(dialogOpts).then(async (returnValue) => {
     if (returnValue.response === 0) {
       log.info('Stretchly: restoring default settings')
       settings.store = Object.assign(require('./utils/defaultSettings'), { isFirstRun: false })
       initialize(false)
-      event.sender.send('renderSettings', settingsToSend())
+      event.sender.send('renderSettings', await settingsToSend())
     }
   })
 })
 
-ipcMain.on('send-settings', function (event) {
-  event.sender.send('renderSettings', settingsToSend())
+ipcMain.on('send-settings', async function (event) {
+  event.sender.send('renderSettings', await settingsToSend())
 })
 
-function settingsToSend () {
-  const loginItemSettings = app.getLoginItemSettings()
-  const openAtLogin = loginItemSettings.openAtLogin
-  return Object.assign({}, settings.store, { openAtLogin: openAtLogin })
+async function settingsToSend () {
+  return Object.assign({}, settings.store, { openAtLogin: await autostartManager.autoLaunchStatus() })
 }
 
 ipcMain.on('play-sound', function (event, sound) {
@@ -1346,12 +1496,16 @@ ipcMain.on('play-sound', function (event, sound) {
 
 ipcMain.on('show-debug', function (event) {
   const reference = breakPlanner.scheduler.reference
-  const timeleft = Utils.formatTimeRemaining(breakPlanner.scheduler.timeLeft)
+  const timeleft = Utils.formatTimeRemaining(breakPlanner.scheduler.timeLeft, settings.get('language'))
   const breaknumber = breakPlanner.breakNumber
   const postponesnumber = breakPlanner.postponesNumber
   const doNotDisturb = breakPlanner.dndManager.isOnDnd
-  const settingsFile = settings.path
-  const logsFile = log.transports.file.getFile().path
+  let settingsFile = settings.path
+  let logsFile = log.transports.file.getFile().path
+  if (process.windowsStore) {
+    settingsFile = settingsFile.replace('Roaming', 'Local\\Packages\\33881JanHovancik.stretchly_24fg4m0zq65je\\LocalCache\\Roaming')
+    logsFile = logsFile.replace('Roaming', 'Local\\Packages\\33881JanHovancik.stretchly_24fg4m0zq65je\\LocalCache\\Roaming')
+  }
   event.sender.send('debugInfo', reference, timeleft,
     breaknumber, postponesnumber, settingsFile, logsFile, doNotDisturb)
 })
@@ -1364,7 +1518,7 @@ ipcMain.on('set-contributor', function (event) {
   const dir = app.getPath('userData')
   const contributorStampFile = `${dir}/stamp`
   const { DateTime } = require('luxon')
-  require('fs').writeFile(contributorStampFile, DateTime.now().toString(), () => {})
+  require('fs').writeFile(contributorStampFile, DateTime.now().toString(), () => { })
   global.shared.isContributor = true
   log.info('Stretchly: Logged in. Thanks for your contributions!')
   if (preferencesWin) {
@@ -1393,7 +1547,8 @@ ipcMain.on('open-contributor-auth', function (event, provider) {
     backgroundColor: 'whitesmoke',
     webPreferences: {
       preload: path.resolve(__dirname, './electron-bridge.js'),
-      enableRemoteModule: true
+      enableRemoteModule: true,
+      sandbox: false
     }
   })
   require('@electron/remote/main').enable(myStretchlyWindow.webContents)
