@@ -4,9 +4,6 @@ const Utils = remote.require('./utils/utils')
 const HtmlTranslate = require('./utils/htmlTranslate')
 const Store = require('electron-store')
 const settings = new Store()
-const log = require('electron-log')
-const path = require('path')
-log.transports.file.resolvePath = () => path.join(remote.app.getPath('userData'), 'logs/main.log')
 
 window.onload = (e) => {
   ipcRenderer.send('send-microbreak-data')
@@ -31,13 +28,12 @@ window.onload = (e) => {
   })
 
   ipcRenderer.once('progress', (event, started, duration, strictMode, postpone, postponePercent, backgroundColor) => {
-    ipcRenderer.send('mini-break-loaded')
     const progress = document.querySelector('#progress')
     const progressTime = document.querySelector('#progress-time')
     const postponeElement = document.querySelector('#postpone')
     const closeElement = document.querySelector('#close')
-    const mainColor = settings.get('mainColor')
-    document.body.classList.add(mainColor.substring(1))
+    const miniBreakColor = settings.get('miniBreakColor')
+    document.body.classList.add(miniBreakColor.substring(1))
     document.body.style.backgroundColor = backgroundColor
 
     document.querySelectorAll('.tiptext').forEach(tt => {
@@ -57,8 +53,10 @@ window.onload = (e) => {
         closeElement.style.display =
           Utils.canSkip(strictMode, postpone, passedPercent, postponePercent) ? 'flex' : 'none'
         progress.value = (100 - passedPercent) * progress.max / 100
-        progressTime.innerHTML = Utils.formatTimeRemaining(Math.trunc(duration - Date.now() + started))
+        progressTime.innerHTML = Utils.formatTimeRemaining(Math.trunc(duration - Date.now() + started),
+          settings.get('language'))
       }
     }, 100)
+    ipcRenderer.send('mini-break-loaded')
   })
 }
