@@ -1,5 +1,5 @@
 const EventEmitter = require('events')
-const log = require('electron-log')
+const log = require('electron-log/main')
 
 class DndManager extends EventEmitter {
   constructor (settings) {
@@ -8,6 +8,19 @@ class DndManager extends EventEmitter {
     this.monitorDnd = settings.get('monitorDnd')
     this.timer = null
     this.isOnDnd = false
+
+    this._unsupDEErrorShown = false
+
+    if (process.platform === 'win32') {
+      this.windowsFocusAssist = require('windows-focus-assist')
+      this.windowsQuietHours = require('windows-quiet-hours')
+    } else if (process.platform === 'darwin') {
+      this.util = require('node:util')
+    } else if (process.platform === 'linux') {
+      this.bus = require('dbus-final').sessionBus()
+      this.util = require('node:util')
+    }
+
     if (this.monitorDnd) {
       this.start()
     }
@@ -17,6 +30,9 @@ class DndManager extends EventEmitter {
     this.monitorDnd = true
     this._checkDnd()
     log.info('Stretchly: starting Do Not Disturb monitoring')
+    if (process.platform === 'linux') {
+      log.info(`System: Your Desktop seems to be ${this._desktopEnviroment}.`)
+    }
   }
 
   stop () {
@@ -27,22 +43,121 @@ class DndManager extends EventEmitter {
     log.info('Stretchly: stopping Do Not Disturb monitoring')
   }
 
-  get _doNotDisturb () {
-    if (this.monitorDnd) {
-      let focusAssist
-      try {
-        focusAssist = require('windows-focus-assist').getFocusAssist().value
-      } catch (e) { focusAssist = -1 } // getFocusAssist() throw an error if OS isn't windows
+  get _desktopEnviroment () {
+    // https://github.com/electron/electron/issues/40795
+    // https://specifications.freedesktop.org/mime-apps-spec/latest/file.html
+    // https://specifications.freedesktop.org/menu-spec/latest/onlyshowin-registry.html
+    return process.env.ORIGINAL_XDG_CURRENT_DESKTOP ||
+      process.env.XDG_CURRENT_DESKTOP || 'unknown'
+  }
 
-      return require('@meetfranz/electron-notification-state').getDoNotDisturb() || (focusAssist !== -1 && focusAssist !== 0)
+  async _isDndEnabledLinux () {
+    const de = this._desktopEnviroment.toLowerCase()
+
+    switch (true) {
+      case de.includes('kde'):
+        try {
+          const obj = await this.bus.getProxyObject('org.freedesktop.Notifications', '/org/freedesktop/Notifications')
+          const properties = obj.getInterface('org.freedesktop.DBus.Properties')
+          const dndEnabled = await properties.Get('org.freedesktop.Notifications', 'Inhibited')
+          if (await dndEnabled.value) {
+            return true
+          }
+        } catch (e) { }
+        break
+      case de.includes('xfce'):
+        try {
+          const obj = await this.bus.getProxyObject('org.xfce.Xfconf', '/org/xfce/Xfconf')
+          const properties = obj.getInterface('org.xfce.Xfconf')
+          const dndEnabled = await properties.GetProperty('xfce4-notifyd', '/do-not-disturb')
+          if (await dndEnabled.value) {
+            return true
+          }
+        } catch (e) { }
+        break
+      case de.includes('gnome') || de.includes('unity'):
+        try {
+          const exec = this.util.promisify(require('node:child_process').exec)
+          const { stdout } = await exec('gsettings get org.gnome.desktop.notifications show-banners')
+          if (stdout.replace(/[^0-9a-zA-Z]/g, '') === 'false') {
+            return true
+          }
+        } catch (e) { }
+        break
+      case de.includes('cinnamon'):
+        try {
+          const exec = this.util.promisify(require('node:child_process').exec)
+          const { stdout } = await exec('gsettings get org.cinnamon.desktop.notifications display-notifications')
+          if (stdout.replace(/[^0-9a-zA-Z]/g, '') === 'false') {
+            return true
+          }
+        } catch (e) { }
+        break
+      case de.includes('mate'):
+        try {
+          const exec = this.util.promisify(require('node:child_process').exec)
+          const { stdout } = await exec('gsettings get org.mate.NotificationDaemon do-not-disturb')
+          if (stdout.replace(/[^0-9a-zA-Z]/g, '') === 'true') {
+            return true
+          }
+        } catch (e) { }
+        break
+      case de.includes('lxqt'):
+        return await this._getConfigValue('~/.config/lxqt/notifications.conf', 'doNotDisturb')
+      default:
+        if (!this._unsupDEErrorShown) {
+          log.info(`Stretchly: ${this._desktopEnviroment} not supported for DND detection, yet.`)
+          this._unsupDEErrorShown = true
+        }
+        return false
+    }
+  }
+
+  async _doNotDisturb () {
+    // TODO also check for session state? https://github.com/felixrieseberg/electron-notification-state/tree/master#session-state
+    if (this.monitorDnd) {
+      if (process.platform === 'win32') {
+        let wfa = 0
+        try {
+          wfa = this.windowsFocusAssist.getFocusAssist().value
+        } catch (e) { wfa = -1 } // getFocusAssist() throw an error if OS isn't windows
+        const wqh = this.windowsQuietHours.getIsQuietHours()
+        return wqh || (wfa !== -1 && wfa !== 0)
+      } else if (process.platform === 'darwin') {
+        try {
+          const exec = this.util.promisify(require('node:child_process').exec)
+          const { stdout } = await exec('defaults read com.apple.controlcenter "NSStatusItem Visible FocusModes"')
+          if (stdout.replace(/[^0-9a-zA-Z]/g, '') === '1') {
+            return true
+          }
+        } catch (e) { }
+      } else if (process.platform === 'linux') {
+        return await this._isDndEnabledLinux()
+      }
     } else {
       return false
     }
   }
 
+  async _getConfigValue (filePath, key) {
+    try {
+      const data = await require('fs').promises.readFile(filePath, 'utf8')
+      const lines = data.split('\n')
+      for (const line of lines) {
+        const [configKey, value] = line.split('=')
+        if (configKey.trim() === key) {
+          return value.trim().toLowerCase() === 'true'
+        }
+      }
+      return false
+    } catch (e) {
+      return false
+    }
+  }
+
   _checkDnd () {
-    this.timer = setInterval(() => {
-      const doNotDisturb = this._doNotDisturb
+    this.timer = setInterval(async () => {
+      const doNotDisturb = await this._doNotDisturb()
       if (!this.isOnDnd && doNotDisturb) {
         this.isOnDnd = true
         this.emit('dndStarted')
